requirements.txt
Django
djangorestframework
djangorestframework_simplejwt
python-dotenv
django-cors-headers
celery
drf-yasg
